feat: 회원 인증 기능 추가

+++ src/main/java/com/lhj/myoppingmall/item/service/ItemService.java
@@ -107,11 +107,14 @@ public class ItemService {
 
     /*
     * 상품 수정
-    * TODO: 권한 인증 추가 필요
     * */
     @Transactional
-    public void updateItem(Long itemId, ItemUpdateRequestDto dto) {
+    public void updateItem(Long itemId, Long memberId, ItemUpdateRequestDto dto) {
         Item item = findItemByItemId(itemId);
+
+        if (!item.getSeller().getId().equals(memberId)) {
+            throw new CustomException(ErrorCode.FORBIDDEN);
+        }
 
         item.changeCommon(dto.getName(), dto.getPrice(), dto.getPictureUrl(), dto.getStockQuantity());
 
@@ -133,11 +136,14 @@ public class ItemService {
 
     /*
     * 상품 삭제
-    * TODO: 권한 인증 추가 필요
     * */
     @Transactional
-    public void deleteItem(Long itemId) {
+    public void deleteItem(Long itemId, Long memberId) {
         Item item = findItemByItemId(itemId);
+
+        if (!item.getSeller().getId().equals(memberId)) {
+            throw new CustomException(ErrorCode.FORBIDDEN);
+        }
 
         itemRepository.delete(item);
     }
